+++ FLEA/FLEA/Db/Exception/MissingLink.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace FLEA\Db\Exception;
+
 
 /**
  * 定义 FLEA_Db_Exception_MissingLink 异常
@@ -16,7 +18,7 @@
  * @author toohamster
  * @version 1.0
  */
-class FLEA_Db_Exception_MissingLink extends FLEA_Exception
+class MissingLink extends \FLEA\Exception
 {
     public $name;